fix(remote-control): when multistream is enabled
