+++ react/features/base/tracks/actionTypes.ts
@@ -10,16 +10,6 @@
 export const SET_NO_SRC_DATA_NOTIFICATION_UID = 'SET_NO_SRC_DATA_NOTIFICATION_UID';
 
 /**
- * The type of redux action dispatched to disable screensharing or to start the
- * flow for enabling screenshare.
- *
- * {
- *     type: TOGGLE_SCREENSHARING
- * }
- */
-export const TOGGLE_SCREENSHARING = 'TOGGLE_SCREENSHARING';
-
-/**
  * The type of redux action dispatched when a track has been (locally or
  * remotely) added to the conference.
  *
