+++ react/features/base/tracks/logger.ts
@@ -1,5 +1,3 @@
-// @flow
-
 import { getLogger } from '../logging/functions';
 
 export default getLogger('features/base/tracks');
