+++ react/features/filmstrip/components/web/ScreenshareFilmstrip.js
@@ -2,8 +2,7 @@
 import React from 'react';
 
 import { connect } from '../../../base/redux';
-import { LAYOUT_CLASSNAMES } from '../../../conference/components/web/Conference';
-import { LAYOUTS, getCurrentLayout } from '../../../video-layout';
+import { LAYOUTS, LAYOUT_CLASSNAMES, getCurrentLayout } from '../../../video-layout';
 import {
     FILMSTRIP_TYPE
 } from '../../constants';
