+++ react/features/filmstrip/components/web/StageFilmstrip.js
@@ -4,8 +4,7 @@ import React from 'react';
 import { getToolbarButtons } from '../../../base/config';
 import { isMobileBrowser } from '../../../base/environment/utils';
 import { connect } from '../../../base/redux';
-import { LAYOUT_CLASSNAMES } from '../../../conference/components/web/Conference';
-import { LAYOUTS, getCurrentLayout } from '../../../video-layout';
+import { LAYOUTS, LAYOUT_CLASSNAMES, getCurrentLayout } from '../../../video-layout';
 import {
     ASPECT_RATIO_BREAKPOINT,
     FILMSTRIP_TYPE,
