+++ react/features/remote-control/actions.js
@@ -5,9 +5,15 @@ import $ from 'jquery';
 import { getMultipleVideoSendingSupportFeatureFlag } from '../base/config/functions.any';
 import { openDialog } from '../base/dialog';
 import { JitsiConferenceEvents } from '../base/lib-jitsi-meet';
-import { getParticipantDisplayName, getPinnedParticipant, pinParticipant } from '../base/participants';
+import {
+    getParticipantDisplayName,
+    getPinnedParticipant,
+    getVirtualScreenshareParticipantByOwnerId,
+    pinParticipant
+} from '../base/participants';
 import { getLocalDesktopTrack, getLocalVideoTrack, toggleScreensharing } from '../base/tracks';
 import { NOTIFICATION_TIMEOUT_TYPE, showNotification } from '../notifications';
+import { isScreenVideoShared } from '../screen-share';
 
 import {
     CAPTURE_EVENTS,
@@ -198,9 +204,12 @@ export function processPermissionRequestReply(participantId: string, event: Obje
                 // the remote control permissions has been granted
                 // pin the controlled participant
                 const pinnedParticipant = getPinnedParticipant(state);
+                const virtualScreenshareParticipantId = getVirtualScreenshareParticipantByOwnerId(state, participantId);
                 const pinnedId = pinnedParticipant?.id;
 
-                if (pinnedId !== participantId) {
+                if (virtualScreenshareParticipantId && pinnedId !== virtualScreenshareParticipantId) {
+                    dispatch(pinParticipant(virtualScreenshareParticipantId));
+                } else if (!virtualScreenshareParticipantId && pinnedId !== participantId) {
                     dispatch(pinParticipant(participantId));
                 }
             }
@@ -508,6 +517,10 @@ export function sendStartRequest() {
         const { sourceId } = track?.jitsiTrack || {};
         const { transport } = state['features/remote-control'].receiver;
 
+        if (typeof sourceId === 'undefined') {
+            return Promise.reject(new Error('Cannot identify screen for the remote control session'));
+        }
+
         return transport.sendRequest({
             name: REMOTE_CONTROL_MESSAGE_NAME,
             type: REQUESTS.start,
@@ -536,7 +549,7 @@ export function grant(participantId: string) {
         const tracks = state['features/base/tracks'];
         const isMultiStreamSupportEnabled = getMultipleVideoSendingSupportFeatureFlag(state);
         const track = isMultiStreamSupportEnabled ? getLocalDesktopTrack(tracks) : getLocalVideoTrack(tracks);
-        const isScreenSharing = track?.videoType === 'desktop';
+        const isScreenSharing = isScreenVideoShared(state);
         const { sourceType } = track?.jitsiTrack || {};
 
         if (isScreenSharing && sourceType === 'screen') {
