+++ react/features/screen-share/components/ShareScreenWarningDialog.tsx
@@ -4,6 +4,8 @@ import { WithTranslation } from 'react-i18next';
 import { IStore } from '../../app/types';
 import { translate } from '../../base/i18n/functions';
 import { connect } from '../../base/redux/functions';
+// eslint-disable-next-line lines-around-comment
+// @ts-ignore
 import { toggleScreensharing } from '../../base/tracks/actions';
 import Dialog from '../../base/ui/components/web/Dialog';
