+++ react/features/video-layout/constants.ts
@@ -9,3 +9,16 @@ export const LAYOUTS = {
     VERTICAL_FILMSTRIP_VIEW: 'vertical-filmstrip-view',
     STAGE_FILMSTRIP_VIEW: 'stage-filmstrip-view'
 };
+
+
+/**
+ * The CSS class to apply so CSS can modify the app layout.
+ *
+ * @private
+ */
+export const LAYOUT_CLASSNAMES = {
+    [LAYOUTS.HORIZONTAL_FILMSTRIP_VIEW]: 'horizontal-filmstrip',
+    [LAYOUTS.TILE_VIEW]: 'tile-view',
+    [LAYOUTS.VERTICAL_FILMSTRIP_VIEW]: 'vertical-filmstrip',
+    [LAYOUTS.STAGE_FILMSTRIP_VIEW]: 'stage-filmstrip'
+};
